# backend/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Customer Data API",
    "description": "API documentation for customer data operations",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/customers": {
      "get": {
        "summary": "Get list of customers",
        "description": "Returns a list of customers with optional filters and pagination",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "customer_name",
            "in": "query",
            "description": "Filter by customer name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer_location",
            "in": "query",
            "description": "Filter by customer location",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "sno": {
                        "type": "integer"
                      },
                      "customer_name": {
                        "type": "string"
                      },
                      "age": {
                        "type": "integer"
                      },
                      "phone": {
                        "type": "string"
                      },
                      "customer_location": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/customer/name/{customer_name}": {
      "get": {
        "summary": "Get customer details by name",
        "description": "Returns customers matching the provided name",
        "parameters": [
          {
            "name": "customer_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "sno": {
                        "type": "integer"
                      },
                      "customer_name": {
                        "type": "string"
                      },
                      "age": {
                        "type": "integer"
                      },
                      "phone": {
                        "type": "string"
                      },
                      "customer_location": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/customer/location/{customer_location}": {
      "get": {
        "summary": "Get customers by location",
        "description": "Returns customers matching the provided location",
        "parameters": [
          {
            "name": "customer_location",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "sno": {
                        "type": "integer"
                      },
                      "customer_name": {
                        "type": "string"
                      },
                      "age": {
                        "type": "integer"
                      },
                      "phone": {
                        "type": "string"
                      },
                      "customer_location": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}